Pass additional user information to hook params. Add new auth handler services and new global hook for authentication.
