+++ src/services/account/hooks/JWTAccountHandlerHooks.js
@@ -1,0 +1,31 @@
+const hooks = require('feathers-hooks-common');
+
+exports.before = {
+	all: [],
+	find: [],
+	get: [hooks.disallow('external')],
+	create: [],
+	update: [],
+	patch: [],
+	remove: [],
+};
+
+exports.after = {
+	all: [],
+	find: [],
+	get: [],
+	create: [],
+	update: [],
+	patch: [],
+	remove: [],
+};
+
+exports.error = {
+	all: [],
+	find: [],
+	get: [],
+	create: [],
+	update: [],
+	patch: [],
+	remove: [],
+};
