+++ src/services/account/hooks/hooks.js
@@ -1,0 +1,232 @@
+const errors = require('@feathersjs/errors');
+const bcrypt = require('bcryptjs');
+const globalHooks = require('../../../hooks');
+
+const MoodleLoginStrategy = require('../strategies/moodle');
+const ITSLearningLoginStrategy = require('../strategies/itslearning');
+const IServLoginStrategy = require('../strategies/iserv');
+const LocalLoginStrategy = require('../strategies/local');
+const LdapLoginStrategy = require('../strategies/ldap');
+
+// don't initialize strategies here - otherwise massive overhead
+// TODO: initialize all strategies here once
+const strategies = {
+	moodle: MoodleLoginStrategy,
+	itslearning: ITSLearningLoginStrategy,
+	iserv: IServLoginStrategy,
+	local: LocalLoginStrategy,
+	ldap: LdapLoginStrategy,
+};
+
+const sanitizeUsername = (hook) => {
+	if (hook.data && hook.data.username && !hook.data.systemId) {
+		hook.data.username = hook.data.username.trim().toLowerCase();
+	}
+	return hook;
+};
+
+// This is only for SSO
+const validateCredentials = (hook) => {
+	const {
+		username, password, systemId, schoolId,
+	} = hook.data;
+
+	if (!username) throw new errors.BadRequest('no username specified');
+	if (!password) throw new errors.BadRequest('no password specified');
+
+	if (!systemId) return hook;
+
+	const { app } = hook;
+	const systemService = app.service('/systems');
+	return systemService.get(systemId)
+		.then((system) => {
+			const Strategy = strategies[system.type];
+			return {
+				strategy: new Strategy(app),
+				system,
+			};
+		})
+		.then(({ strategy, system }) => strategy.login({ username, password }, system, schoolId))
+		.then((client) => {
+			if (client.token) {
+				hook.data.token = client.token;
+			}
+			return hook;
+		});
+};
+
+const trimPassword = (hook) => {
+	if (hook.data.password) {
+		hook.data.password = hook.data.password.trim();
+	}
+	if (hook.data.password_verification) {
+		hook.data.password_verification = hook.data.password_verification.trim();
+	}
+
+	return hook;
+};
+
+const validatePassword = (hook) => {
+	const passwordVerification = hook.data.password_verification;
+	const { password } = hook.data;
+
+	// Check against Pattern which is also used in Frontend
+	const pattern = new RegExp('^(?=.*[A-Z])(?=.*[0-9])(?=.*[a-z])(?=.*[\\-_!<>§$%&\\/()=?\\\\;:,.#+*~\']).{8,255}$');
+	const patternResult = pattern.test(password);
+
+	// only check result if also a password was really given
+	if (!patternResult && password) {
+		throw new errors.BadRequest('Dein Passwort stimmt mit dem Pattern nicht überein.');
+	}
+
+	// in case sso created account skip
+	if (!hook.params.account.userId) {
+		return hook;
+	}
+
+	return Promise.all([
+		globalHooks.hasPermissionNoHook(hook, hook.params.account.userId, 'STUDENT_CREATE'),
+		globalHooks.hasRoleNoHook(hook, hook.id, 'student', true),
+		globalHooks.hasPermissionNoHook(hook, hook.params.account.userId, 'ADMIN_VIEW'),
+		globalHooks.hasRoleNoHook(hook, hook.id, 'teacher', true),
+		globalHooks.hasRole(hook, hook.params.account.userId, 'superhero')])
+		.then(([hasStudentCreate, isStudent, hasAdminView, isTeacher, isSuperHero]) => {
+			const editsOwnAccount = (hook.params.account._id || {}).toString() === hook.id;
+			if (
+				(hasStudentCreate && isStudent)
+				|| (hasAdminView && (isStudent || isTeacher))
+				|| isSuperHero
+				|| editsOwnAccount) {
+				return hook;
+			}
+			if (password && !passwordVerification) {
+				throw new errors.Forbidden(
+					`Du darfst das Passwort dieses Nutzers nicht ändern oder die
+					Passwortfelder wurden falsch ausgefüllt.`,
+				);
+			}
+
+			if (passwordVerification) {
+				return new Promise((resolve, reject) => {
+					bcrypt.compare(passwordVerification, hook.params.account.password, (err, res) => {
+						if (err) {
+							reject(new errors.BadRequest('Ups, bcrypt ran into an error.'));
+						}
+						if (!res) {
+							reject(new errors.BadRequest('Dein Passwort ist nicht korrekt!'));
+						}
+						resolve();
+					});
+				});
+			}
+			return hook;
+		});
+};
+
+const checkUnique = (hook) => {
+	const accountService = hook.service;
+	const { username, systemId } = hook.data;
+	return accountService.find({ query: { username, systemId } })
+		.then((result) => {
+			// systemId might be null. In that case, accounts with any systemId will be returned
+			const filtered = result.filter(a => a.systemId === systemId);
+			if (filtered.length > 0) {
+				return Promise.reject(new errors.BadRequest('Der Benutzername ist bereits vergeben!'));
+			}
+			return Promise.resolve(hook);
+		});
+};
+
+const removePassword = (hook) => {
+	const { strategy } = hook.data;
+	if (strategy === 'ldap') {
+		hook.data.password = '';
+	}
+	return Promise.resolve(hook);
+};
+
+const restrictAccess = (hook) => {
+	const queries = hook.params.query;
+
+	return new Promise((resolve, reject) => {
+		if (!queries.username && !queries.userId) {
+			return reject(new errors.BadRequest('Not allowed'));
+		} return resolve();
+	});
+};
+
+const checkExistence = (hook) => {
+	const accountService = hook.service;
+	const { userId, systemId } = hook.data;
+
+	if (!userId && systemId) { // for sso accounts
+		return Promise.resolve(hook);
+	}
+
+	return accountService.find({ query: { userId } })
+		.then((result) => {
+			// systemId might be null. In that case, accounts with any systemId will be returned
+			const filtered = result.filter(a => a.systemId === systemId);
+			if (filtered.length > 0) return Promise.reject(new errors.BadRequest('Der Account existiert bereits!'));
+			return Promise.resolve(hook);
+		});
+};
+
+const protectUserId = (hook) => {
+	const accountService = hook.service;
+	if (hook.data.userId) {
+		return accountService.get(hook.id)
+			.then((res) => {
+				if (res.systemId) {
+					return Promise.resolve(hook);
+				} return Promise.reject(new errors.Forbidden('Die userId kann nicht geändert werden.'));
+			});
+	}
+	return hook;
+};
+
+const securePatching = hook => Promise.all([
+	globalHooks.hasRole(hook, hook.params.account.userId, 'superhero'),
+	globalHooks.hasRole(hook, hook.params.account.userId, 'administrator'),
+	globalHooks.hasRole(hook, hook.params.account.userId, 'teacher'),
+	globalHooks.hasRoleNoHook(hook, hook.id, 'student', true),
+]).then(([isSuperHero, isAdmin, isTeacher, targetIsStudent]) => {
+	const editsOwnAccount = (hook.params.account._id || {}).toString() === hook.id;
+	if (hook.params.account._id !== hook.id) {
+		if (!(isSuperHero || isAdmin || (isTeacher && targetIsStudent) || editsOwnAccount)) {
+			return Promise.reject(new errors.BadRequest('You have not the permissions to change other users'));
+		}
+	}
+	return Promise.resolve(hook);
+});
+
+/**
+ * @method get
+ * @param {Array of strings} keys
+ * @afterHook
+ * @notLocal
+ */
+const filterToRelated = keys => globalHooks.ifNotLocal((hook) => {
+	const newResult = {};
+	keys.forEach((key) => {
+		if (hook.result[key] !== undefined) {
+			newResult[key] = hook.result[key];
+		}
+	});
+	hook.result = newResult;
+	return hook;
+});
+
+module.exports = {
+	validateCredentials,
+	trimPassword,
+	validatePassword,
+	checkUnique,
+	removePassword,
+	restrictAccess,
+	checkExistence,
+	protectUserId,
+	securePatching,
+	filterToRelated,
+	sanitizeUsername,
+};
