+++ src/services/account/hooksCJWT/index.js
@@ -1,4 +1,3 @@
-const auth = require('@feathersjs/authentication');
 const hooks = require('feathers-hooks-common');
 const globalHooks = require('../../../hooks');
 
@@ -6,7 +5,7 @@ exports.before = {
 	all: [],
 	find: [hooks.disallow()],
 	get: [hooks.disallow()],
-	create: [auth.hooks.authenticate('jwt'), globalHooks.isSuperHero()],
+	create: [globalHooks.authenticateJWT, globalHooks.isSuperHero()],
 	update: [hooks.disallow()],
 	patch: [hooks.disallow()],
 	remove: [hooks.disallow()],
