+++ src/services/account/model.js
@@ -1,23 +1,22 @@
-'use strict';
-
 const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
+
+const { Schema } = mongoose;
 
 const accountSchema = new Schema({
-	username: {type: String, required: true, lowercase: true},
-	password: {type: String},
+	username: { type: String, required: true, lowercase: true },
+	password: { type: String },
 
-	token: {type: String},
-	credentialHash: {type: String},
+	token: { type: String },
+	credentialHash: { type: String },
 
-	userId: {type: Schema.Types.ObjectId, ref: 'user'},
-	systemId: {type: Schema.Types.ObjectId, ref: 'system'}, // if systemId => SSO
+	userId: { type: Schema.Types.ObjectId, ref: 'user' },
+	systemId: { type: Schema.Types.ObjectId, ref: 'system' }, // if systemId => SSO
 
-	expiresAt: {type: Date},
+	expiresAt: { type: Date },
 
-	activated: {type: Boolean, 'default': false}
-},{
-	timestamps: true
+	activated: { type: Boolean, default: false },
+}, {
+	timestamps: true,
 });
 
 const accountModel = mongoose.model('account', accountSchema);
