+++ src/services/account/services/index.js
@@ -1,0 +1,5 @@
+const JWTAccountHandler = require('./JWTAccountHandler');
+
+module.exports = {
+	JWTAccountHandler,
+};
