+++ src/services/analytics/hooks/index.js
@@ -1,8 +1,9 @@
 const auth = require('@feathersjs/authentication');
+const globalHooks = require('../../../hooks');
 
 exports.before = {
 	all: [
-		auth.hooks.authenticate('jwt'),
+		globalHooks.authenticateJWT,
 	],
 	find: [],
 	get: [],
