+++ src/services/federalState/hooks/index.js
@@ -1,8 +1,7 @@
-const auth = require('@feathersjs/authentication');
 const globalHooks = require('../../../hooks');
 
 exports.before = {
-	all: [auth.hooks.authenticate('jwt')],
+	all: [globalHooks.authenticateJWT],
 	find: [globalHooks.hasPermission('FEDERALSTATE_VIEW')],
 	get: [globalHooks.hasPermission('FEDERALSTATE_VIEW')],
 	create: [globalHooks.hasPermission('FEDERALSTATE_CREATE')],
