+++ src/services/homework/hooks/comments.js
@@ -1,4 +1,3 @@
-const auth = require('@feathersjs/authentication');
 const globalHooks = require('../../../hooks');
 
 
@@ -10,7 +9,7 @@ const addToSubmission = (hook) => {
 };
 
 exports.before = () => ({
-	all: [auth.hooks.authenticate('jwt')],
+	all: [globalHooks.authenticateJWT],
 	find: [globalHooks.hasPermission('COMMENTS_VIEW'), globalHooks.mapPaginationQuery.bind(this)],
 	get: [globalHooks.hasPermission('COMMENTS_VIEW')],
 	create: [globalHooks.hasPermission('COMMENTS_CREATE')],
