+++ src/services/homework/hooks/copy.js
@@ -1,5 +1,4 @@
 const hooks = require('feathers-hooks-common');
-const auth = require('@feathersjs/authentication');
 const errors = require('@feathersjs/errors');
 const globalHooks = require('../../../hooks');
 const HomeworkModel = require('../model').homeworkModel;
@@ -18,7 +17,7 @@ const hasViewPermissionBefore = (hook) => {
 };
 
 exports.before = {
-	all: [auth.hooks.authenticate('jwt')],
+	all: [globalHooks.authenticateJWT],
 	find: [hooks.disallow()],
 	get: [
 		globalHooks.hasPermission('HOMEWORK_VIEW'),
