+++ src/services/passwordRecovery/hooks/index.js
@@ -1,6 +1,3 @@
-'use strict';
-
-const auth = require('@feathersjs/authentication');
 const local = require('@feathersjs/authentication-local');
 const globalHooks = require('../../../hooks');
 
@@ -23,7 +20,7 @@ const hashId = (hook) => {
 exports.before = {
 	all: [],
 	find: [
-		auth.hooks.authenticate('jwt'),
+		globalHooks.authenticateJWT,
 		globalHooks.hasPermission('PWRECOVERY_VIEW'),
 	],
 	get: [],
@@ -32,16 +29,16 @@ exports.before = {
 		local.hooks.hashPassword({ passwordField: 'password' }),
 	],
 	update: [
-		auth.hooks.authenticate('jwt'),
+		globalHooks.authenticateJWT,
 		globalHooks.hasPermission('PWRECOVERY_EDIT'),
 	],
 	patch: [
-		auth.hooks.authenticate('jwt'),
+		globalHooks.authenticateJWT,
 		globalHooks.hasPermission('PWRECOVERY_EDIT'),
 		globalHooks.permitGroupOperation,
 	],
 	remove: [
-		auth.hooks.authenticate('jwt'),
+		globalHooks.authenticateJWT,
 		globalHooks.hasPermission('PWRECOVERY_CREATE'),
 		globalHooks.permitGroupOperation,
 	],
