+++ src/services/system/hooks/index.js
@@ -1,4 +1,3 @@
-const auth = require('@feathersjs/authentication');
 const { permitGroupOperation } = require('../../../hooks');
 const globalHooks = require('../../../hooks');
 
@@ -6,10 +5,10 @@ exports.before = {
 	all: [],
 	find: [],
 	get: [],
-	create: [auth.hooks.authenticate('jwt'), globalHooks.hasPermission('SYSTEM_CREATE')],
-	update: [auth.hooks.authenticate('jwt'), globalHooks.hasPermission('SYSTEM_EDIT')],
-	patch: [auth.hooks.authenticate('jwt'), globalHooks.hasPermission('SYSTEM_EDIT'), permitGroupOperation],
-	remove: [auth.hooks.authenticate('jwt'), globalHooks.hasPermission('SYSTEM_CREATE'), permitGroupOperation],
+	create: [globalHooks.authenticateJWT, globalHooks.hasPermission('SYSTEM_CREATE')],
+	update: [globalHooks.authenticateJWT, globalHooks.hasPermission('SYSTEM_EDIT')],
+	patch: [globalHooks.authenticateJWT, globalHooks.hasPermission('SYSTEM_EDIT'), permitGroupOperation],
+	remove: [globalHooks.authenticateJWT, globalHooks.hasPermission('SYSTEM_CREATE'), permitGroupOperation],
 };
 
 exports.after = {
