+++ src/services/user-group/hooks/copyCourseHook.js
@@ -1,5 +1,4 @@
 const hooks = require('feathers-hooks-common');
-const auth = require('@feathersjs/authentication');
 const _ = require('lodash');
 const globalHooks = require('../../../hooks');
 
@@ -10,7 +9,7 @@ const injectCourseId = (hook) => {
 };
 
 exports.before = {
-	all: [auth.hooks.authenticate('jwt'), globalHooks.hasPermission('USERGROUP_CREATE')],
+	all: [globalHooks.authenticateJWT, globalHooks.hasPermission('USERGROUP_CREATE')],
 	find: [hooks.disallow()],
 	get: [hooks.disallow()],
 	create: [globalHooks.injectUserId, injectCourseId, globalHooks.ifNotLocal(globalHooks.checkCorrectCourseOrTeamId)],
@@ -30,7 +29,7 @@ exports.after = {
 };
 
 exports.beforeShare = {
-	all: [auth.hooks.authenticate('jwt'), globalHooks.hasPermission('USERGROUP_CREATE')],
+	all: [globalHooks.authenticateJWT, globalHooks.hasPermission('USERGROUP_CREATE')],
 	find: [],
 	get: [globalHooks.ifNotLocal(globalHooks.checkCorrectCourseOrTeamId)],
 	create: [globalHooks.injectUserId],
