+++ src/services/user/hooks/firstLogin.js
@@ -1,11 +1,12 @@
 const hooks = require('feathers-hooks-common');
-const auth = require('@feathersjs/authentication');
+
+const globalHooks = require('../../../hooks');
 
 exports.before = {
 	all: [],
 	find: hooks.disallow('external'),
 	get: hooks.disallow('external'),
-	create: [auth.hooks.authenticate('jwt')],
+	create: [globalHooks.authenticateJWT],
 	update: hooks.disallow('external'),
 	patch: hooks.disallow('external'),
 	remove: hooks.disallow('external'),
