+++ src/services/wopi/hooks/index.js
@@ -1,7 +1,8 @@
 /* eslint-disable no-multi-spaces */
 const { BadRequest, Conflict, NotFound } = require('@feathersjs/errors');
-const auth = require('@feathersjs/authentication');
+
 const { FileModel } = require('../../fileStorage/model');
+const globalHooks = require('../../../hooks');
 
 /**
 	* handles the authentication for wopi-clients,
@@ -28,7 +29,7 @@ const wopiAuthentication = (hook) => {
 		jwt = jwt.slice(0, jwt.indexOf('?permission'));
 	}
 	hook.params.headers.authorization = jwt.replace('Bearer ', '');
-	return auth.hooks.authenticate('jwt')(hook);
+	return globalHooks.authenticateJWT(hook);
 };
 /**
  * All editing (POST, PATCH, DELETE) actions should include the wopi-override header!
